requirements.txt
Django==5.0.6
djangorestframework==3.15.2
numpy==1.26.4
scipy==1.13.1
librosa==0.10.2.post1
soundfile==0.12.1
pydub==0.25.1
tensorflow-cpu==2.20.0
tensorflow-hub==0.16.1
tflite-support==0.4.3
noisereduce==3.0.0
py-webrtcvad==2.0.10